Adding a Jar upload to workflow
Allow for a binary artifact of a .jar file of the game to be donwloaded. This may have the same issues as the JaCoCo action with an unreconised file path. If this is the case, then is is the file paths to the build.gradle that are the problem and not JaCoCo itself.