Helping .jar file find main class
The jar file, when executed, can't find the main class and hence cannot run. (Likely due to it being in desktop and not core). As such I have updated the main class from project.mainClassName (which I think is just a default and not specific to our project) to com.team13.piazzapanic.DesktopLauncher, which is the package + class containing the main class. I've also added 'class-path', which should add the class path to the Manifest.mf, allowing the main class to be found.