Adding Pan test
Adding a test for the update class as it overrides the Cooking Station class. Also, slightly amended the cooking station class as I left in a print statement and removed an assertion as it was unnecessary. (Checked for the class Failed Ingredient not the boolean value, which cannot be checked for a null ingredient).

Signed-off-by: CrimsonLeaves <109231078+CrimsonLeaves@users.noreply.github.com>